requirements.txt
flask
requests
python-magic
pyelftools